Fix `tileconv` paths for Linux & Mac
The mod uses `HANDLE_TILECONV`, which in turn checks for the tileconv
executable in a subfolder named after the content of the `%WEIDU_OS%`
variable.

That variable holds `osx` and `unix`, respectively, on Mac and Linux -
not `macosx` and `linux. See:
http://www.weidu.org/~thebigg/README-WeiDU.html#WEIDU_OS

This commit renames those two subfolders accordingly.

I've tested that this change makes installation of the mod succeed on
Linux (whereas it didn't before).